Rename political_commentry files to political_commentry_finder
