+++ agents/political_commentry_finder.py
@@ -11,7 +11,7 @@ available commentary on local issues.
 from dotenv import load_dotenv
 
 from agents.base_agent_template import BaseReActAgent
-from tools.political_commentry import (
+from tools.political_commentry_finder import (
     political_figure_finder,
     search_political_commentary,
 )
