+++ tools/political_commentry_finder.py
